+++ jdaviz/configs/cubeviz/plugins/slice/slice.py
@@ -1,6 +1,9 @@
+import threading
+import time
 import warnings
-from astropy.units import UnitsWarning
+
 import numpy as np
+from astropy.units import UnitsWarning
 from glue_jupyter.bqplot.image import BqplotImageView
 from glue_jupyter.bqplot.profile import BqplotProfileView
 from traitlets import Bool, Float, observe, Any, Int
@@ -45,6 +48,9 @@ class Slice(PluginTemplateMixin):
     show_indicator = Bool(True).tag(sync=True)
     show_wavelength = Bool(True).tag(sync=True)
 
+    is_playing = Bool(False).tag(sync=True)
+    play_interval = Int(200).tag(sync=True)  # milliseconds
+
     def __init__(self, *args, **kwargs):
         self._default_spectrum_viewer_reference_name = kwargs.get(
             "spectrum_viewer_reference_name", "spectrum-viewer"
@@ -57,6 +63,7 @@ class Slice(PluginTemplateMixin):
         self._watched_viewers = []
         self._indicator_viewers = []
         self._x_all = None
+        self._player = None
 
         # initialize watching existing viewers WITH data (if initializing the plugin after data
         # already exists - otherwise the AddDataMessage will handle watching image viewers once
@@ -173,7 +180,10 @@ class Slice(PluginTemplateMixin):
 
     @observe('slice')
     def _on_slider_updated(self, event):
-        value = int(event.get('new', int(len(self._x_all)/2)))
+        if self._x_all is None:
+            return
+
+        value = int(event.get('new', int(len(self._x_all)/2))) % (int(self.max_value) + 1)
 
         self.wavelength = self._x_all[value]
 
@@ -181,3 +191,41 @@ class Slice(PluginTemplateMixin):
             viewer.state.slices = (0, 0, value)
         for viewer in self._indicator_viewers:
             viewer._update_slice_indicator(value)
+
+    def vue_goto_first(self, *args):
+        if self.is_playing:
+            return
+        self._on_slider_updated({'new': self.min_value})
+
+    def vue_goto_last(self, *args):
+        if self.is_playing:
+            return
+        self._on_slider_updated({'new': self.max_value})
+
+    def vue_play_next(self, *args):
+        if self.is_playing:
+            return
+        self._on_slider_updated({'new': self.slice + 1})
+
+    def _player_worker(self):
+        ts = float(self.play_interval) * 1e-3  # ms to s
+        while self.is_playing:
+            self._on_slider_updated({'new': self.slice + 1})
+            time.sleep(ts)
+
+    def vue_play_start_stop(self, *args):
+        if self.is_playing:  # Stop
+            if self._player:
+                if self._player.is_alive():
+                    self._player.join(timeout=0)
+                self._player = None
+            self.is_playing = False
+            return
+
+        if self._x_all is None:
+            return
+
+        # Start
+        self.is_playing = True
+        self._player = threading.Thread(target=self._player_worker)
+        self._player.start()